requirements.txt
Django>=4.2.3
Pillow>=9.5.0
django-widget-tweaks>=1.4.8
django-allauth>=0.65.7
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=42.0.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1 